pyproject.toml
[project]
name = "alerta"
version = "0.2.0"
description = "An Alert engine for the defendA data lake"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.72",
    "chevron>=0.14.0",
    "netaddr>=1.3.0",
    "pandas>=3.0.1",
    "pymongo>=4.16.0",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.3",
    "tzlocal>=5.3.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-docker-tools>=3.1.9",
]
